remove dependency on aws rds in masstransit core
