--- src/MassTransit/SqlTransport/Configuration/AccessToken/ITokenGenerator.cs
@@ -1,6 +1,0 @@
-namespace MassTransit.AccessToken;
-
-public interface ITokenGenerator
-{
-    string GenerateToken(SqlTransportOptions options);
-}
