+++ src/Transports/MassTransit.SqlTransport.PostgreSql/Configuration/AccessToken/ITokenGenerator.cs
@@ -1,0 +1,9 @@
+namespace MassTransit.AccessToken
+{
+    public interface ITokenGenerator
+    {
+        string GenerateToken(SqlTransportOptions options);
+    }
+
+}
+
